requirements.txt
streamlit>=1.28.0
pycryptodome>=3.19.0
qrcode[pil]>=7.4.0
Pillow>=10.0.0
